Module aws_sdk_networkmanager::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AssociateCustomerGateway

Fluent builder constructing a request to AssociateCustomerGateway.

AssociateLink

Fluent builder constructing a request to AssociateLink.

AssociateTransitGatewayConnectPeer

Fluent builder constructing a request to AssociateTransitGatewayConnectPeer.

CreateConnection

Fluent builder constructing a request to CreateConnection.

CreateDevice

Fluent builder constructing a request to CreateDevice.

CreateGlobalNetwork

Fluent builder constructing a request to CreateGlobalNetwork.

CreateLink

Fluent builder constructing a request to CreateLink.

CreateSite

Fluent builder constructing a request to CreateSite.

DeleteConnection

Fluent builder constructing a request to DeleteConnection.

DeleteDevice

Fluent builder constructing a request to DeleteDevice.

DeleteGlobalNetwork

Fluent builder constructing a request to DeleteGlobalNetwork.

DeleteLink

Fluent builder constructing a request to DeleteLink.

DeleteSite

Fluent builder constructing a request to DeleteSite.

DeregisterTransitGateway

Fluent builder constructing a request to DeregisterTransitGateway.

DescribeGlobalNetworks

Fluent builder constructing a request to DescribeGlobalNetworks.

DisassociateCustomerGateway

Fluent builder constructing a request to DisassociateCustomerGateway.

DisassociateLink

Fluent builder constructing a request to DisassociateLink.

DisassociateTransitGatewayConnectPeer

Fluent builder constructing a request to DisassociateTransitGatewayConnectPeer.

GetConnections

Fluent builder constructing a request to GetConnections.

GetCustomerGatewayAssociations

Fluent builder constructing a request to GetCustomerGatewayAssociations.

GetDevices

Fluent builder constructing a request to GetDevices.

GetLinkAssociations

Fluent builder constructing a request to GetLinkAssociations.

GetLinks

Fluent builder constructing a request to GetLinks.

GetNetworkResourceCounts

Fluent builder constructing a request to GetNetworkResourceCounts.

GetNetworkResourceRelationships

Fluent builder constructing a request to GetNetworkResourceRelationships.

GetNetworkResources

Fluent builder constructing a request to GetNetworkResources.

GetNetworkRoutes

Fluent builder constructing a request to GetNetworkRoutes.

GetNetworkTelemetry

Fluent builder constructing a request to GetNetworkTelemetry.

GetRouteAnalysis

Fluent builder constructing a request to GetRouteAnalysis.

GetSites

Fluent builder constructing a request to GetSites.

GetTransitGatewayConnectPeerAssociations

Fluent builder constructing a request to GetTransitGatewayConnectPeerAssociations.

GetTransitGatewayRegistrations

Fluent builder constructing a request to GetTransitGatewayRegistrations.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

RegisterTransitGateway

Fluent builder constructing a request to RegisterTransitGateway.

StartRouteAnalysis

Fluent builder constructing a request to StartRouteAnalysis.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateConnection

Fluent builder constructing a request to UpdateConnection.

UpdateDevice

Fluent builder constructing a request to UpdateDevice.

UpdateGlobalNetwork

Fluent builder constructing a request to UpdateGlobalNetwork.

UpdateLink

Fluent builder constructing a request to UpdateLink.

UpdateNetworkResourceMetadata

Fluent builder constructing a request to UpdateNetworkResourceMetadata.

UpdateSite

Fluent builder constructing a request to UpdateSite.