Module aws_sdk_networkmanager::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AssociateCustomerGatewayError

AssociateLinkError

AssociateTransitGatewayConnectPeerError

ConflictException

There was a conflict processing the request. Updating or deleting the resource can

CreateConnectionError

CreateDeviceError

CreateGlobalNetworkError

CreateLinkError

CreateSiteError

DeleteConnectionError

DeleteDeviceError

DeleteGlobalNetworkError

DeleteLinkError

DeleteSiteError

DeregisterTransitGatewayError

DescribeGlobalNetworksError

DisassociateCustomerGatewayError

DisassociateLinkError

DisassociateTransitGatewayConnectPeerError

GetConnectionsError

GetCustomerGatewayAssociationsError

GetDevicesError

GetLinkAssociationsError

GetLinksError

GetSitesError

GetTransitGatewayConnectPeerAssociationsError

GetTransitGatewayRegistrationsError

InternalServerException

The request has failed due to an internal error.

ListTagsForResourceError

RegisterTransitGatewayError

ResourceNotFoundException

The specified resource could not be found.

ServiceQuotaExceededException

A service limit was exceeded.

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateConnectionError

UpdateDeviceError

UpdateGlobalNetworkError

UpdateLinkError

UpdateSiteError

ValidationException

The input fails to satisfy the constraints.

Enums

AssociateCustomerGatewayErrorKind

AssociateLinkErrorKind

AssociateTransitGatewayConnectPeerErrorKind

CreateConnectionErrorKind

CreateDeviceErrorKind

CreateGlobalNetworkErrorKind

CreateLinkErrorKind

CreateSiteErrorKind

DeleteConnectionErrorKind

DeleteDeviceErrorKind

DeleteGlobalNetworkErrorKind

DeleteLinkErrorKind

DeleteSiteErrorKind

DeregisterTransitGatewayErrorKind

DescribeGlobalNetworksErrorKind

DisassociateCustomerGatewayErrorKind

DisassociateLinkErrorKind

DisassociateTransitGatewayConnectPeerErrorKind

GetConnectionsErrorKind

GetCustomerGatewayAssociationsErrorKind

GetDevicesErrorKind

GetLinkAssociationsErrorKind

GetLinksErrorKind

GetSitesErrorKind

GetTransitGatewayConnectPeerAssociationsErrorKind

GetTransitGatewayRegistrationsErrorKind

ListTagsForResourceErrorKind

RegisterTransitGatewayErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateConnectionErrorKind

UpdateDeviceErrorKind

UpdateGlobalNetworkErrorKind

UpdateLinkErrorKind

UpdateSiteErrorKind