Module aws_sdk_networkmanager::model

Modules

aws_location

See AwsLocation

bandwidth

See Bandwidth

connection

See Connection

customer_gateway_association

See CustomerGatewayAssociation

device

See Device

global_network

See GlobalNetwork

link

See Link

link_association

See LinkAssociation

location

See Location

site

See Site

tag

See Tag

transit_gateway_connect_peer_association

See TransitGatewayConnectPeerAssociation

transit_gateway_registration

See TransitGatewayRegistration

transit_gateway_registration_state_reason

See TransitGatewayRegistrationStateReason

validation_exception_field

See ValidationExceptionField

Structs

AwsLocation

Specifies a location in AWS.

Bandwidth

Describes bandwidth information.

Connection

Describes a connection.

CustomerGatewayAssociation

Describes the association between a customer gateway, a device, and a link.

Device

Describes a device.

GlobalNetwork

Describes a global network.

Link

Describes a link.

LinkAssociation

Describes the association between a device and a link.

Location

Describes a location.

Site

Describes a site.

Tag

Describes a tag.

TransitGatewayConnectPeerAssociation

Describes a transit gateway Connect peer association.

TransitGatewayRegistration

Describes the registration of a transit gateway to a global network.

TransitGatewayRegistrationStateReason

Describes the status of a transit gateway registration.

ValidationExceptionField

Describes a validation exception for a field.

Enums

ConnectionState

CustomerGatewayAssociationState

DeviceState

GlobalNetworkState

LinkAssociationState

LinkState

SiteState

TransitGatewayConnectPeerAssociationState

TransitGatewayRegistrationState

ValidationExceptionReason