Module aws_sdk_networkmanager::operation

Structs

AssociateCustomerGateway

Associates a customer gateway with a device and optionally, with a link. If you

AssociateLink

Associates a link to a device. A device can be associated to multiple links and a link can be associated to multiple devices. The device and link must be in the same global network and the same site.

AssociateTransitGatewayConnectPeer

Associates a transit gateway Connect peer with a device, and optionally, with a link. If you

CreateConnection

Creates a connection between two devices. The devices can be a physical or virtual appliance that connects to a third-party appliance in a VPC, or a physical appliance that connects to another physical appliance in an on-premises network.

CreateDevice

Creates a new device in a global network. If you specify both a site ID and a

CreateGlobalNetwork

Creates a new, empty global network.

CreateLink

Creates a new link for a specified site.

CreateSite

Creates a new site in a global network.

DeleteConnection

Deletes the specified connection in your global network.

DeleteDevice

Deletes an existing device. You must first disassociate the device from any links and

DeleteGlobalNetwork

Deletes an existing global network. You must first delete all global network objects

DeleteLink

Deletes an existing link. You must first disassociate the link from any devices and

DeleteSite

Deletes an existing site. The site cannot be associated with any device or link.

DeregisterTransitGateway

Deregisters a transit gateway from your global network. This action does not delete

DescribeGlobalNetworks

Describes one or more global networks. By default, all global networks are

DisassociateCustomerGateway

Disassociates a customer gateway from a device and a link.

DisassociateLink

Disassociates an existing device from a link. You must first disassociate any customer

DisassociateTransitGatewayConnectPeer

Disassociates a transit gateway Connect peer from a device and link.

GetConnections

Gets information about one or more of your connections in a global network.

GetCustomerGatewayAssociations

Gets the association information for customer gateways that are associated with

GetDevices

Gets information about one or more of your devices in a global network.

GetLinkAssociations

Gets the link associations for a device or a link. Either the device ID or the link ID

GetLinks

Gets information about one or more links in a specified global network.

GetSites

Gets information about one or more of your sites in a global network.

GetTransitGatewayConnectPeerAssociations

Gets information about one or more of your transit gateway Connect peer associations in a global network.

GetTransitGatewayRegistrations

Gets information about the transit gateway registrations in a specified

ListTagsForResource

Lists the tags for a specified resource.

RegisterTransitGateway

Registers a transit gateway in your global network. The transit gateway can be in any

TagResource

Tags a specified resource.

UntagResource

Removes tags from a specified resource.

UpdateConnection

Updates the information for an existing connection. To remove information for any of the parameters,

UpdateDevice

Updates the details for an existing device. To remove information for any of the

UpdateGlobalNetwork

Updates an existing global network. To remove information for any of the parameters,

UpdateLink

Updates the details for an existing link. To remove information for any of the

UpdateSite

Updates the information for an existing site. To remove information for any of the