Module aws_sdk_networkmanager::model

Data structures used by operation inputs/outputs.

Modules

aws_location

See AwsLocation

bandwidth

See Bandwidth

connection

See Connection

connection_health

See ConnectionHealth

customer_gateway_association

See CustomerGatewayAssociation

device

See Device

global_network

See GlobalNetwork

link

See Link

link_association

See LinkAssociation

location

See Location

network_resource

See NetworkResource

network_resource_count

See NetworkResourceCount

network_resource_summary

See NetworkResourceSummary

network_route

See NetworkRoute

network_route_destination

See NetworkRouteDestination

network_telemetry

See NetworkTelemetry

path_component

See PathComponent

relationship

See Relationship

route_analysis

See RouteAnalysis

route_analysis_completion

See RouteAnalysisCompletion

route_analysis_endpoint_options

See RouteAnalysisEndpointOptions

route_analysis_endpoint_options_specification

See RouteAnalysisEndpointOptionsSpecification

route_analysis_path

See RouteAnalysisPath

route_table_identifier

See RouteTableIdentifier

site

See Site

tag

See Tag

transit_gateway_connect_peer_association

See TransitGatewayConnectPeerAssociation

transit_gateway_registration

See TransitGatewayRegistration

transit_gateway_registration_state_reason

See TransitGatewayRegistrationStateReason

validation_exception_field

See ValidationExceptionField

Structs

AwsLocation

Specifies a location in Amazon Web Services.

Bandwidth

Describes bandwidth information.

Connection

Describes a connection.

ConnectionHealth

Describes connection health.

CustomerGatewayAssociation

Describes the association between a customer gateway, a device, and a link.

Device

Describes a device.

GlobalNetwork

Describes a global network.

Link

Describes a link.

LinkAssociation

Describes the association between a device and a link.

Location

Describes a location.

NetworkResource

Describes a network resource.

NetworkResourceCount

Describes a resource count.

NetworkResourceSummary

Describes a network resource.

NetworkRoute

Describes a network route.

NetworkRouteDestination

Describes the destination of a network route.

NetworkTelemetry

Describes the telemetry information for a resource.

PathComponent

Describes a path component.

Relationship

Describes a resource relationship.

RouteAnalysis

Describes a route analysis.

RouteAnalysisCompletion

Describes the status of an analysis at completion.

RouteAnalysisEndpointOptions

Describes a source or a destination.

RouteAnalysisEndpointOptionsSpecification

Describes a source or a destination.

RouteAnalysisPath

Describes a route analysis path.

RouteTableIdentifier

Describes a route table.

Site

Describes a site.

Tag

Describes a tag.

TransitGatewayConnectPeerAssociation

Describes a transit gateway Connect peer association.

TransitGatewayRegistration

Describes the registration of a transit gateway to a global network.

TransitGatewayRegistrationStateReason

Describes the status of a transit gateway registration.

ValidationExceptionField

Describes a validation exception for a field.

Enums

ConnectionState

ConnectionStatus

ConnectionType

CustomerGatewayAssociationState

DeviceState

GlobalNetworkState

LinkAssociationState

LinkState

RouteAnalysisCompletionReasonCode

RouteAnalysisCompletionResultCode

RouteAnalysisStatus

RouteState

RouteTableType

RouteType

SiteState

TransitGatewayConnectPeerAssociationState

TransitGatewayRegistrationState

ValidationExceptionReason